Keep the Adam step counter in FP32 under a BF16 default dtype
_AdamBase.step() initialized the step counter as torch.tensor(0.0), which picks
up torch.get_default_dtype() instead of being pinned to FP32. Under
torch.set_default_dtype(torch.bfloat16) the counter is BF16, which cannot
represent 257, so `state["step"] += 1` stops advancing at 256.

bias_correction1/bias_correction2 are derived from that frozen step, so
bias_correction2 stays pinned at 1 - 0.999**256 = 0.226 instead of converging to
1. denom = sqrt(v) / sqrt(bias_correction2) is then permanently ~2.1x too large
and every update after step 256 is silently too small. With a constant gradient
and lr=1e-3, a parameter that should reach ~1.0e-5 after 1000 steps instead sits
at 0.204.

Upstream guards against this: torch.optim.optimizer._get_scalar_dtype() returns
FP32 regardless of the default dtype, and torch.optim.AdamW keeps counting
correctly in the same configuration. The same class already pins lr to FP32 in
add_param_group.

test_optim_default_dtype_bf16 runs a single step, far below the saturation point,
so it could not catch this. Add a test that asserts the step counter's dtype and
that it advances past 256.

+++ torchao/optim/adam.py
@@ -121,7 +121,11 @@ class _AdamBase(Optimizer):
 
                     # State initialization
                     if len(state) == 0:
-                        state["step"] = torch.tensor(0.0)
+                        # Pin the step counter to FP32 instead of letting it pick up the
+                        # default dtype. Under torch.set_default_dtype(torch.bfloat16) the
+                        # counter would be BF16, which cannot represent 257, so it saturates
+                        # at 256 and the bias corrections below freeze with it.
+                        state["step"] = torch.tensor(0.0, dtype=torch.float32)
                         state["exp_avg"] = self._new_buffer(p, True)
                         state["exp_avg_sq"] = self._new_buffer(p, False)
                         if group["amsgrad"]:
